This change addresses the following two action items:
http://www.w3.org/2011/webappsec/track/actions/44
http://www.w3.org/2011/webappsec/track/actions/48

Specifically, rather than including the HTTP request line in violation reports,
we now include the full document-uri (sans fragment) and the document's
referrer.

Previously, the spec was inconsistent about whether the user agent should
re-write any occurances of 'self' in the policy with the actual value of
'self'.  Folks on the mailing list didn't have strong opinions, so I've
resolved the inconsistency in favor of not re-writing.