Converting "allowed XXX sources" from `var` to `dfn` elements.
This has the effect of making them linkable outside of their individual
sections, which will be of use when defining the script interface methods
in an upcoming patch.

As a driveby, this patch also corrects a reference to "allowed font sources"
within the "connect-src" section, and a reference to "script-types" in the
"plugin-types" section.