CSP 1.1: Loosening the 'script-nonce' grammar.
The 'script-nonce' directive's value is currently defined by the grammar laid
out for "token" in section 2.2 of RFC 2616. This is unnecessarily restrictive;
there's no good reason that '@', '/', and other separators shouldn't be valued
members of the noncing community.

This patch changes the grammar from "token" to "VCHAR minus ',' and ';'", which
better suits the directive's requirements and implementation.