CSP 1.1: Changing 'plugin-types' behavior to strictly enforce content types.
As proposed on the list[1], this patch changes the specification to require
that authors making use of the 'plugin-types' directive explicitly declare the
expected media type of resources loaded via 'object' and 'embed' elements. If
a 'type' attribute isn't present, or its value doesn't match the content type of
the loaded resource, then user agents must reject the resource, and should
display fallback content.

This new behavior is experimentally implemented in WebKit[2].

[1]: http://lists.w3.org/Archives/Public/public-webappsec/2012Aug/0000.html
[2]: http://webk.it/91919