CSP 1.1: Cleaning up 'sandbox' usage section; no normative changes.
The original text lead to unintended conclusions about the 'sandbox' directive's
effect on things like plugins. This patch changes that text to remove the poorly
phrased example in favor of more explicit descriptions of the scenarios I had in
mind.

Additionally, it adds a note about the defense-in-depth nature of the directive.
I'm not sure this is necessary, as that's implicit in the whole document, but
Devdatta Akhawe suggested it (thanks!), and it certainly doesn't hurt to repeat
important points every so often.