CSP 1.1: Expanding the path-matching syntax to include files.
After some discussion on the list[1], it seems reasonable to add file-level
granularity to the source expression matching rules. This patch is a first
stab at that, with the intent of allowing source expressions like
'example.com/js/file.js' to match specific files.

In short, paths that end in '/' will be treated as directories, matching all
files under that directory. Paths that do not end in '/' will be considered
files, matching only one specific entry.