CSP 1.1: First pass at adding line numbers to reports.
Based on public-webappsec conversation[1], there seems to be rough agreement
that adding a source/line number to violation reports would be useful for
developers as they try to track down those reports' sources. Firefox already
sends this information in reports, and has done so for years.

This patch is a first pass for discussion, simply bringing in the keys and
values Firefox is already using. It makes the new fields optional, as not
every report will have this sort of context available, and user agents may
choose not to send this information in certain circumstances (extensions,
addons, etc).

[1]: http://lists.w3.org/Archives/public-webappsec/2012Dec/0025.html