CSP 1.1: Add non-normative examples of enforcing multiple policies.
The question of multiple-policy enforcement came up in December twice on the
list. The proper handling is technically clear in section 3.1.1/3.1.2, but could
be much clearer to developers with an example.

This patch adds section 3.1.4 with an example to attempt to address the deficit
of clarity.

This should close ACTION-106.