CSP 1.1: Non-normative note about the limits of the script interface.
The script interface as currently defined by CSP 1.1 only gives insight into
the state of the page's CSP. It does not promise that allowed origins are, in
fact, _accessible_. CORS, same-origin policy, etc. will all place additional
limitations on requests above and beyond CSP. Now that's explicit in the spec.