CSP 1.1: When reporting blocked 'data:' resources, send 'data:' as the URL.
Based on Twitter's feedback from real-world error reports[1], it would be useful to
clarify the expected behavior for resources that don't map well to the general
'http:'/'https:' schemes we expect. If a 'data:...' resource is blocked, the
spec will now direct implementors to report 'data:'. Likewise, blob resources
would report 'blob:', JavaScript resources would report 'javascript:' and so on.

[1]: http://lists.w3.org/Archives/Public/public-webappsec/2013Feb/0011.html