CSP 1.1: Strawman SecurityPolicyViolationEvent interface.
This patch adds a strawman SecurityPolicyViolationEvent interface, and wires it
up to fire events on the protected resource's Document node on every violation.
In the interests of simplicity, it currently replicates exactly the data
available in existing violation reports. There's been some discussion around
three types of changes that I'll stub out in future patches:

1. Policies should be able to opt-out of DOM-level reporting, as discussed in
   a November 2012 thread[1].

2. Violation events should possibly include a DOM reference (perhaps as a
   'relatedTarget'?) to the node that triggered the error (see the same
   thread[1] for context).

3. A running list of violation events should potentially be stored
   ('document.securityPolicyViolations'?) so that violations that occur before
   a developer hooks up an event handler can still be dealt with in some
   fashion.

[1]: http://lists.w3.org/Archives/Public/public-webappsec/2012Nov/0128.html