CSP 1.1: Initialize SecurityPolicyViolationEvent attributes.
We now explicitly initialize the attributes of SecurityPolicyViolationEvent
objects as part of "fire a violation event", and the MUST requirement to
actually fire the event has moved to the "Processing Model" section for
clarity. Thanks again to Anne for pointing out this deficency.

One important change to note: the values now differ slightly from the values
in the POSTed violation report: we're no longer stripping fragments or in any
other way adjusting the URLs. The reason is simply that JavaScript will already
have access to that data: stripping the fragment identifier on the document
when it's trivially accessible via 'window.location.hash' seems like more of an
annoyance than a security protection.