Trim SecurityPolicy interface
As discussed at the WebAppSec face-to-face meeting, there aren't strong use
cases for all the detailed SecurityPolicy functions and there's some risk that
exposing the set of report URIs could leak some information.

This change trims the SecurityPolicy interface to just the first four
attributes, which the folks at the face-to-face agreed have strong use cases.