CSP 1.1: Queue a task to deliver violation reports.
The spec should make it clear that the user agent is in control of violation
report delivery. Here, we add an explicit task source for Content Security
Policy, and ensure that violation reports are queued up in such a way that
enables the user agent to pick and choose exactly when reports are delivered.