CSP 1.1: A first-pass at defining behavior for script/style hashes.
This patch builds off the work Neil Matatall, Garrett Robinson, Joel
Weinberger, and others have done to outline the contours of a hashing
algorithm that CSP can support. In particular, [1] and [2] are highly
relevant background reading.

Most of the heavy lifting with regard to determining the exact text
that ought to be hashed is kicked out to the HTML spec, which helpfully
defines "the script block's source" in a way that I hope will clarify
the exact algorithm necessary in order to ensure interoperability between
user agents when processing text.

Some items to note:

* This patch includes support for only SHA-256, SHA-384, and SHA-512
  digest algorithms, as defined by the NIST. We might want to include
  older algorithms (SHA-1?) as well, and it's been suggested that we
  include non-NIST algorithms. These are topics for future patches.

* This patch allows the 'unsafe-inline' source to continue working
  alongside hashes and nonces. There are good reasons to believe that
  we should change this behavior to ignore 'unsafe-inline' when either
  nonces or hashes are present. I'll make that change in a future patch
  so that it's its own topic of discussion.

[1]: http://lists.w3.org/Archives/Public/public-webappsec/2013Aug/0031.html
[2]: http://www.w3.org/2011/webappsec/minutes/webappsec-minutes-13-Aug-2013.html