CSP 1.1: Disable 'unsafe-inline' if a hash or nonce are present.
Nonces and hashes should obviate the need for 'unsafe-inline'. Several
threads ([1] and [2] most notably) have suggested options for dealing
with this scenario by disabling 'unsafe-inline' iff a nonce or hash
source is present, as this would allow forward-thinking site authors
to whitelist specific inline scripts in browsers that support CSP 1.1,
without breaking support for all inline scripts in browsers that
support CSP 1.0.

For obvious reasons, we should encourage these sites (and browser vendors)
to move as rapidly as is reasonable to CSP 1.1 support.

[1]: http://lists.w3.org/Archives/Public/public-webappsec/2013Jul/0019.html
[2]: http://lists.w3.org/Archives/Public/public-webappsec/2013Dec/0004.html