CSP 1.1: Remove the script interface to the set of enforced policies.
The current API is poor, and not what we'd like to run with going
forward. Alex Russell and Yehuda Katz have strongly argued for and
outlined a better take on an API ([1], [2]). Removing the current
definition to make room for a future definition along the lines of
their proposals.

This patch does not, however, remove the event fired upon violation
of the set of currently active policies. That seems to stand alone
quite cleanly.

[1]: http://infrequently.org/2013/05/use-case-zero/
[2]: http://yehudakatz.com/2013/05/24/an-extensible-approach-to-browser-security-policy/