CSP 1.1: Workers may have their own policies.
According to [1], we have consensus that web workers ought not always
inherit their owner documents' set of policies, but instead be governed
by whatever policy is delivered along with the worker's script.

Workers should inherit their owner document's set of policies only when
they are created from URLs with unique origins (such as the various
types of non-hierarchical URLs we care about: blob, filesystem, data,
etc.).

This patch attempts to clarify that, and also clarifies that
'importScripts' ought to be filtered through the currently active set of
policies.

[1]: http://lists.w3.org/Archives/Public/public-webappsec/2013Dec/0007.html