SRI: Allow non-eligible resources with empty integrity attributes
Using the original steps, a resource with an empty integrity
attribute might be blocked if it's not eligible (e.g. non-CORS and
not same-origin). This is wrong since we're not supposed to do
any checks or block anything in the "no metadata" case.

With this change, empty integrity attributes will always pass the
checks and never be blocked for SRI reasons.