SECURE: Define the worker relationship correctly.
As @bzbarsky noted WAY back in June[1], the worker text makes no sense
as it refers to a property which doesn't exist. This patch cleans that
up a bit by walking through the Worker's Documents, which is a real thing,
and not the responsible Document, which is not.

This should improve the handling of Shared Workers too, which is nice.

[1]: https://lists.w3.org/Archives/Public/public-webappsec/2015Jul/0010.html